Add rudimentary CMake build system
Currently can build tsmuxer CLI and libmediation in native 64-bit
on Linux.  Windows, Mac, and 32-bit untested (32-bit is mostly
a function of the user's CXX_FLAGS settings anyway).

As a consequence of trying to get it to build the correct files,
the textSubtitleRender{FT|Win32}.* files had to be moved to an
osdep/ subdirectory.  Some #include cases were also fixed to
point to the correct headers and directories in the source
tree or elsewhere.

Install rules need to be fleshed out for static/shared libs and
headers for libmediation, uninstall rules need to be added.

tsMuxerGUI needs to be hooked in as well.