Use standard-width types in convertUTF.h
convertUTF.h typedefed the UTF32 type as unsigned long, which is really 64-bit
wide on LP64 platforms. Since there is a number of casts, mostly around
utf8Converter.cpp, which assume that UTF32 is 32 bits and no more, this caused
some of the pointers to point outside of their respective arrays due to pointer
arithmetic adding an offset of 8 instead of 4.

This naturally never occured on Windows, because it uses 32 bits for longs, and
the problematic code was never actually really hit because sizeof(wchar_t) on
Windows is 2, which means that the code which needed to cast to UTF32 was never
executed.

Fixes #69.