Use the checked out Git revision as app version if building in a Git repository
This replaces the occurences of the static version number with a version string
if building in a Git repository and Git is available on the build host.

This isn't perfect, since the revision is only going to be fetched when running
qmake or cmake, while it's possible that it changes between consecutive
compilations, not just reconfigurations. However, the main purpose of this is
to add a version tag to the builds built centrally by the GitHub runners, which
always configure the project from scratch.