UDF 2.50: omit second (N-256) AVDP
UDF 2.50 specification [6.10.1 Use of UDF on CD-R media] states: "ECMA 167 requires an Anchor Volume Descriptor Pointer (AVDP) at sector 256 and either N or (N - 256), where N is the last recorded Physical Address on the media. UDF requires that the AVDP be recorded at both sector 256 and sector (N - 256) when each session is closed."

As both ImgBurn and DVDFab do not incude the (N - 256) Anchor Volume Decritpor, it seems preferable to omit this descriptor for compatibility with medias and players.